测试通过

+++ YTS.IOFile.API.Test/TestPathRuleParsingJSON.cs
@@ -5,12 +5,9 @@ using System.IO;
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-using YTS.IOFile.API.Tools;
-using YTS.IOFile.API.Tools.DataFileIO;
 using YTS.IOFile.API.Tools.PathRuleParsing;
-
-using YTS.Git;
 using YTS.Logic.Log;
+
 using Newtonsoft.Json;
 
 namespace YTS.IOFile.API.Test
@@ -36,19 +33,22 @@ namespace YTS.IOFile.API.Test
         [TestMethod]
         public void Test_ToWrite()
         {
-            const string root_dire = @"C:\_code_test\dire";
+            const string root_dire = @"C:\_code_test";
             // 清理测试环境
-            Directory.Delete(root_dire, true);
+            if (Directory.Exists(root_dire))
+            {
+                Directory.Delete(root_dire, true);
+            }
             ruleParsing.SetRoot(root_dire);
 
             PathResolutionResult result;
 
             result = ruleParsing.ToWrite(@"plan:list");
-            Assert.AreEqual(@"C:\_code_test\dire\plan\list.json", result.AbsolutePathAddress);
+            Assert.AreEqual(@"C:\_code_test\_data\plan\list.json", result.AbsolutePathAddress);
 
-            ruleParsing.SetRoot(@"C:\_code_test\dire\");
+            ruleParsing.SetRoot(@"C:\_code_test\");
             result = ruleParsing.ToWrite(@"plan:list");
-            Assert.AreEqual(@"C:\_code_test\dire\plan\list.json", result.AbsolutePathAddress);
+            Assert.AreEqual(@"C:\_code_test\_data\plan\list.json", result.AbsolutePathAddress);
 
             Directory.Delete(root_dire, true);
         }
@@ -59,10 +59,12 @@ namespace YTS.IOFile.API.Test
             IList<PathResolutionResult> list;
             PathResolutionResult result;
 
-
-            const string root_dire = @"C:\_code_test\dire";
+            const string root_dire = @"C:\_code_test";
             // 清理测试环境
-            Directory.Delete(root_dire, true);
+            if (Directory.Exists(root_dire))
+            {
+                Directory.Delete(root_dire, true);
+            }
             ruleParsing.SetRoot(root_dire);
 
             // 开始获取
@@ -71,28 +73,28 @@ namespace YTS.IOFile.API.Test
 
             // 写入内容
             string json = JsonConvert.SerializeObject(new { });
-            File.WriteAllText(@"C:\_code_test\dire\plan\list.json", json);
+            FileExtend.WriteAllText(@"C:\_code_test\_data\plan\list.json", json);
 
             // 读取地址
             list = ruleParsing.ToRead(@"plan:list");
             Assert.AreEqual(1, list.Count);
             result = list[0];
             Assert.AreEqual(@"plan:list", result.Key);
-            Assert.AreEqual(@"C:\_code_test\dire\plan\list.json", result.AbsolutePathAddress);
+            Assert.AreEqual(@"C:\_code_test\_data\plan\list.json", result.AbsolutePathAddress);
 
             // 写入第二个文件
-            File.WriteAllText(@"C:\_code_test\dire\zwang\list.json", json);
+            FileExtend.WriteAllText(@"C:\_code_test\_data\zwang\list.json", json);
 
             // 再次读取
             list = ruleParsing.ToRead(@"/\w+/i:list");
             Assert.AreEqual(2, list.Count);
             result = list[0];
             Assert.AreEqual(@"plan:list", result.Key);
-            Assert.AreEqual(@"C:\_code_test\dire\plan\list.json", result.AbsolutePathAddress);
+            Assert.AreEqual(@"C:\_code_test\_data\plan\list.json", result.AbsolutePathAddress);
 
             result = list[1];
             Assert.AreEqual(@"zwang:list", result.Key);
-            Assert.AreEqual(@"C:\_code_test\dire\zwang\list.json", result.AbsolutePathAddress);
+            Assert.AreEqual(@"C:\_code_test\_data\zwang\list.json", result.AbsolutePathAddress);
 
             Directory.Delete(root_dire, true);
         }
